Disable web asset caching
While the Unity player seems to try and fall back to not using IndexedDB
for asset caches if it’s unavailable, in Chromium, it is unsuccessful in
doing so and stops loading. IndexedDB is unavailable on Itch for many
people because Itch serves the game in a CDN-delivered iframe, which
means IndexedDB availability depends on third-party cookie settings.